backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
pandas
numpy
scikit-learn
transformers
torch
requests
aiofiles
redis
httpx
phishing-detection-py